Fix strict-overflow bugs and reenable this warning.
In two places, we were counting things in an array using an 'int'.  In
theory, the int could overflow, so gcc determines this to be undefined
behaviour.

The fix is to use size_t or ssize_t instead.

+++ format/format.c
@@ -424,7 +424,7 @@ static int
 do_rescan (char **devices)
 {
   size_t i;
-  int errors = 0;
+  size_t errors = 0;
   guestfs_error_handler_cb old_error_cb;
   void *old_error_data;
